requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-PyMongo
pymongo
python-dateutil
streamlit